pyproject.toml
[project]
name = "fastapi-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.0",
    "google-auth>=2.43.0",
    "google-genai>=1.52.0",
    "openai>=2.8.1",
    "openai-agents>=0.6.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
